requirements.txt
django==4.0
djangorestframework==3.13.1
dj-rest-auth==2.1.11
django-allauth==0.47.0
pyyaml==6.0
uritemplate==4.1.1
drf-yasg==1.20.0